ある程度できたので、push

+++ src/main/java/com/example/blog/repository/UserRepository.java
@@ -16,5 +16,5 @@ public interface UserRepository {
     void update(User user);
 
     @Delete("DELETE FROM users WHERE id = #{id}")
-    void delete(Long id);
+    void delete(Integer id);
 }
